Make localeswitcher accessible
This change ensures the li contains a button which is clicked, rather than adding a role="button" to the li (that role is not allowed on an li and the necessary js to take over all functionality of a button was missing).  It also adds a small bit of JS to the eventhandler for the button to ensure that the lang on the html element is changed when pressing the button.  Without that JS a reload of the page was necessary to ensure the lang attribute had the correct value after clicking a button.

Signed-off-by: Koen Cornelis <koen@ibuildings.nl>